+++ CHANGELOG.md
@@ -141,6 +141,10 @@ The following is a curated list of changes in the Enact limestone module, newest
 
 - `limestone/Slider` styling to match the latest GUI
 
+### Added
+
+- `limestone/TabLayout` `selectOnFocus` prop to enable selection when a tab is focused via pointer mode.
+
 ### Fixed
 
 - `limestone/Card` press motion to work with key press